+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillEqualize.java
@@ -88,11 +88,11 @@ public class SkillEqualize extends TargettedSkill {
         }
         Player tPlayer = (Player) target;
         if (player.getHealth() > tPlayer.getHealth()) {
-            plugin.getServer().getPluginManager().callEvent(new EntityRegainHealthEvent(target, (int) (player.getHealth() - tPlayer.getHealth()), RegainReason.CUSTOM));
+            plugin.getServer().getPluginManager().callEvent(new EntityRegainHealthEvent(target, (player.getHealth() - tPlayer.getHealth()), RegainReason.CUSTOM));
         } else {
-            plugin.getServer().getPluginManager().callEvent(new EntityRegainHealthEvent(player, (int) (tPlayer.getHealth() - player.getHealth()), RegainReason.CUSTOM));
+            plugin.getServer().getPluginManager().callEvent(new EntityRegainHealthEvent(player, (tPlayer.getHealth() - player.getHealth()), RegainReason.CUSTOM));
         }
-        
+        //TODO: er...you do realize that this doesn't actually heal the hero right? just fyi.
         broadcastExecuteText(hero, target);
         return SkillResult.NORMAL;
     }
