+++ src/com/herocraftonline/heroes/characters/skill/skills/SkillTackle.java
@@ -107,7 +107,7 @@ public class SkillTackle extends TargettedSkill {
         long duration = (long) ((SkillConfigManager.getUseSetting(hero, this, SkillSetting.DURATION.node(), 3000, false) +
                 (SkillConfigManager.getUseSetting(hero, this, "duration-increase", 0.0, false) * hero.getSkillLevel(this))));
         duration = duration > 0 ? duration : 0;
-        int damage = (int) (SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE.node(), 6, false) +
+        double damage = (SkillConfigManager.getUseSetting(hero, this, SkillSetting.DAMAGE.node(), 6, false) +
                 (SkillConfigManager.getUseSetting(hero, this, "damage-increase", 0.0, false) * hero.getSkillLevel(this)));
         damage = damage > 0 ? damage : 0;
         for (Entity e : ((Entity) hero.getPlayer()).getNearbyEntities(radius, radius, radius)) {
